GenAI/AgenticAI/edaAgent/requirements.txt
langchain
langchain-experimental
langchain-groq
pandas
python-dotenv
tabulate
